Split Legend into LegendaryHero and Capital
